--- frontend/src/components/RankingDialog.tsx/index.tsx
@@ -1,1 +1,0 @@
-export * from "./RankingDialog";
