+++ frontend/src/components/RatingDialog.tsx/index.tsx
@@ -1,0 +1,1 @@
+export * from "./RatingDialog";
